# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API Recicle Link",
    "description": "A RecicleLink é um projeto desenvolvido durante o hackathon \"Hack for Change\", uma iniciativa realizada em parceria entre a FIAP e a Alura. Este projeto tem como principal objetivo promover o recolhimento adequado dos resíduos sólidos, conectando cooperativas de reciclagem com os coletores, tornando todo o processo mais eficiente e, consequentemente, diminuindo os impactos ambientais causados pelo descarte inadequado dos resíduos naquela região.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "description": "API DEV",
      "url": "http://localhost:5000/api/"
    }
  ],
  "paths": {
    "/usuarios": {
      "get": {
        "summary": "Obtem todos os usuários",
        "responses": {
          "200": {
            "$ref": "#/components/responses/usuarios"
          },
          "500": {
            "$ref": "#/components/responses/erro"
          }
        },
        "tags": ["Usuário"]
      },
      "post": {
        "summary": "Cria um usuário novo",
        "requestBody": {
          "$ref": "#/components/requestBody/usuario"
        },
        "responses": {
          "201": {
            "description": "Sucesso"
          },
          "400": {
            "$ref": "#/components/responses/erroRegistroExiste"
          },
          "500": {
            "$ref": "#/components/responses/erro"
          }
        },
        "tags": ["Usuário"]
      }
    },
    "/usuarios/id": {
      "parameters": [{"$ref": "#/components/parameters/id_string"}],
      "get": {
        "summary": "Obtem usuario por id",
        "responses": {
          "200": {
            "$ref": "#/components/responses/usuario"
          },
          "500": {
            "$ref": "#/components/responses/erro"
          }
        },
        "tags": ["Usuário"]
      }
    },
    "/usuarios/tipoUsuario": {
      "parameters": [{"$ref": "#/components/parameters/tipoUsuario"}],
      "get": {
        "summary": "Obtem usuarios por tipo",
        "responses": {
          "200": {
            "$ref": "#/components/responses/usuarios"
          },
          "500": {
            "$ref": "#/components/responses/erro"
          }
        },
        "tags": ["Usuário"]
      }
    },
    "/usuarios/login": {
      "post": {
        "summary": "Login do usuário",
        "requestBody": {
          "$ref": "#/components/requestBody/usuario_login"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/usuario_token"
          },
          "400": {
            "$ref": "#/components/responses/erroCredencial"
          }
        },
        "tags": ["Usuário"]
      }
    }
  },
  "components": {
    "parameters": {
      "tipoUsuario": {
        "name": "tipoUsuario",
        "description": "(usuario, coletador, cooperativa)",
        "in": "query",
        "schema": {
          "type": "string"
        },
        "required": false
      },
      "id_string": {
        "name": "id",
        "in": "query",
        "schema": {
          "type": "string"
        }
      }
    },
    "requestBody": {
      "usuario": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/usuario_create"
            }
          }
        }
      },
      "usuario_login": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/usuario_login"
            }
          }
        }
      }
    },
    "responses": {
      "usuario": {
        "description": "Sucesso",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/usuario"
            }
          }
        }
      },
      "usuarios": {
        "description": "Sucesso",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/usuarios"
            }
          }
        }
      },
      "usuario_token": {
        "description": "Sucesso",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/usuario_token"
            }
          }
        }
      },
      "erroCredencial": {
        "description": "Erro de credencial inválida",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/error"
            }
          }
        }
      },
      "erroRegistroExiste": {
        "description": "Erro de registro já existente",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/error"
            }
          }
        }
      },
      "erro": {
        "description": "Erro do servidor",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/error"
            }
          }
        }
      }
    },
    "schemas": {
      "usuario": {
        "title": "usuario schema",
        "description": "Usuário (Object)",
        "type": "object",
        "properties": {
          "_id": {
            "title": "Identificação única do usuário",
            "type": "string"
          },
          "nome": {
            "title": "Nome do usuário",
            "type": "string"
          },
          "email": {
            "title": "Email do usuário",
            "type": "string"
          },
          "senha": {
            "title": "Senha criptografada do usuário",
            "type": "string"
          },
          "endereco": {
            "title": "Endereço do usuário",
            "type": "string"
          },
          "tipoUsuario": {
            "title": "Tipo do usuário",
            "type": "string"
          },
          "solicitacoesAtribuidas": {
            "title": "Solicitações atribuída do usuário",
            "type": "string"
          },
          "__v": {
            "title": "Coluna de controle de versão",
            "type": "integer"
          }
        }
      },
      "usuario_create": {
        "title": "schema de criação do usuário",
        "description": "Usuário (Object)",
        "type": "object",
        "properties": {
          "nome": {
            "title": "Nome do usuário",
            "type": "string"
          },
          "email": {
            "title": "Email do usuário",
            "type": "string"
          },
          "senha": {
            "title": "Senha do usuário",
            "type": "string"
          },
          "endereco": {
            "title": "Endereço do usuário",
            "type": "string"
          },
          "tipoUsuario": {
            "title": "Tipo do usuário",
            "type": "string"
          }
        }
      },
      "usuario_login": {
        "title": "schema de login do usuário",
        "description": "Usuário (Object)",
        "type": "object",
        "properties": {
          "email": {
            "title": "Email do usuário",
            "type": "string"
          },
          "senha": {
            "title": "Senha do usuário",
            "type": "string"
          }
        }
      },
      "usuario_token": {
        "title": "schema do token de login do usuário",
        "description": "Token (Object)",
        "type": "object",
        "properties": {
          "token": {
            "title": "Token do usuário",
            "type": "string"
          },
          "idUsuario": {
            "title": "ID do usuário",
            "type": "string"
          },
          "tipoUsuario": {
            "title": "Tipo do usuário",
            "type": "string"
          },
          "solicitacoesAtribuidas": {
            "title": "solicitações atribuidas ao usuário",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "usuarios": {
        "title": "usuarios schema",
        "description": "Usuários (Array)",
        "type": "array",
        "items": {
          "$ref": "#components/schemas/usuario"
        }
      },
      "error": {
        "title": "Erro",
        "description": "Erro (Object)",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  }
}